Setup Custom Cog Support
Added support for custom cogs to be added.
