+++ core/resources/model_settings.csv
@@ -1,0 +1,1 @@
+sdxl|2|DPM++ SDE Karras|1
